Feat: added a bond scraper + serializing a extracted bonds data to json file (#1)
The idea behind this was to create a scraper which will extract bond data from https://www.obligacjeskarbowe.pl/oferta/ and then serialize this to a json file. The bonds offer changes once per year so it's not big deal that it is run manually, pros of this solution it that after extracting data there can be some adjustment done manualny in a json file.